Modularize project management renderer screens
