+++ src/renderer/modules/projektverwaltung/README.md
@@ -1,0 +1,12 @@
+# Projektverwaltung
+
+Dieses Modul bündelt die Renderer-Screens für die bestehende Projektverwaltung.
+
+Enthaltene Screens:
+- `ProjectsScreen`
+- `ProjectFormScreen`
+- `ArchiveScreen`
+
+Wichtig:
+- Keine DB- oder IPC-Logik im Renderer-Modul.
+- Die Projektverwaltung ist der erste fachliche Modul-Meilenstein nach dem eingefrorenen Protokoll-Modul.
