+++ src/renderer/modules/projektverwaltung/index.js
@@ -1,0 +1,1 @@
+export * from "./screens/index.js";
